move EventsServer to GrpcEventWatcher.cs
