+++ sdk/Pulumi.Automation/Events/GrpcEventWatcher.cs
@@ -5,6 +5,10 @@ using System.Linq;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
+using Google.Protobuf.WellKnownTypes;
+using Grpc.Core;
+using Pulumi.Automation.Serialization;
+using Pulumirpc;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Hosting.Server;
@@ -116,6 +120,50 @@ namespace Pulumi.Automation.Events
             _shutdownCts.Dispose();
             _host.Dispose();
             await Task.CompletedTask;
+        }
+    }
+
+    /// <summary>
+    /// gRPC service implementation for receiving engine events over gRPC.
+    /// This service receives events streamed from the Pulumi CLI during stack operations.
+    /// </summary>
+    internal sealed class EventsServer : Pulumirpc.Events.EventsBase
+    {
+        private readonly LocalSerializer _localSerializer = new LocalSerializer();
+        private readonly Action<EngineEvent> _onEvent;
+
+        public EventsServer(Action<EngineEvent> onEvent)
+        {
+            _onEvent = onEvent;
+        }
+
+        public override async Task<Empty> StreamEvents(IAsyncStreamReader<EventRequest> requestStream, ServerCallContext context)
+        {
+            try
+            {
+                while (await requestStream.MoveNext(context.CancellationToken).ConfigureAwait(false))
+                {
+                    var request = requestStream.Current;
+                    var eventJson = request.Event;
+
+                    if (!string.IsNullOrWhiteSpace(eventJson) && _localSerializer.IsValidJson(eventJson))
+                    {
+                        var engineEvent = _localSerializer.DeserializeJson<EngineEvent>(eventJson);
+                        _onEvent.Invoke(engineEvent);
+                    }
+                }
+            }
+            catch (Exception) when (context.CancellationToken.IsCancellationRequested)
+            {
+                // Operation was cancelled, this is expected
+            }
+            catch (Exception)
+            {
+                // Log or handle other exceptions if needed
+                throw;
+            }
+
+            return new Empty();
         }
     }
 }
